Implement a player data handler with JSON
This is essentially a 1:1 copy from the prototype repository.

Player Data instances are managed by the Player Data Manager (obviously).
The instantiation of new data instances and retrieval of instances should be done through it.
In actuality, it delegates load requests to the PlayerDataStorage passed to it in the constructor.

Also implements custom exceptions for singleton instantiation and retrieval.