👷 Perform gradle wizardry
Painstakingly and with the help of ChatGPT, I can finally present
to you the black magic of Ascendancy build configuration.

There are 3 projects included with this commit: API, Core and
BuiltInPack.

Each one of these depends on paper while the later two also depend
on the API project.
Core also depends on Citizens, Sentinel and Carbon.

All dependency versions are managed in `libs.version.toml` under
the gradle/ directory.

The global project version is in `gradle.properties`